Optimized and improved performance
