+++ package.json
@@ -15,7 +15,7 @@
     "@mui/icons-material": "^6.1.9",
     "@mui/material": "^6.1.9",
     "@reduxjs/toolkit": "^2.4.0",
-    "axios": "^1.7.9",
+    "lodash": "^4.17.21",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
