+++ src/components/TableHeadComponent.jsx
@@ -1,0 +1,19 @@
+import { TableHead, TableRow, TableCell } from "@mui/material";
+
+export default function TableHeadComponent({ columns }) {
+  return (
+    <TableHead>
+      <TableRow>
+        {columns.map((column) => (
+          <TableCell
+            key={column.id}
+            align={column.align}
+            style={{ minWidth: column.minWidth }}
+          >
+            {column.label}
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+}
